Refractor item_list_view.dart and split code into multiple file for better redability
